# Conflicts:
#	protocol/v1/qbft/controller/controller.go
#	protocol/v1/qbft/controller/controller_decided.go
#	protocol/v1/qbft/controller/controller_decided_test.go
#	protocol/v1/qbft/controller/controller_running_instance.go
#	protocol/v1/qbft/controller/msgqueue_consumer_test.go
#	protocol/v1/qbft/controller/proccess_message.go
#	protocol/v1/validator/validator.go

+++ network/discovery/local_service.go
@@ -91,13 +91,13 @@ func (md *localDiscovery) FindPeers(ctx context.Context, ns string, opt ...disco
 }
 
 // RegisterSubnets implements Service
-func (md *localDiscovery) RegisterSubnets(subnets ...int64) error {
+func (md *localDiscovery) RegisterSubnets(subnets ...int) error {
 	// TODO
 	return nil
 }
 
 // DeregisterSubnets implements Service
-func (md *localDiscovery) DeregisterSubnets(subnets ...int64) error {
+func (md *localDiscovery) DeregisterSubnets(subnets ...int) error {
 	// TODO
 	return nil
 }
